requirements.txt
discord.py-self==2.0.0
nltk==3.8.1
openai==0.27.6
python-dotenv==1.0.0
